[site] Generate complete DIF description
Currently the text produced for the DIF description stops at the first
subelement. This is apparent if the description contains a variable
enclosed in backticks. The text up to this variable is returned but not
the rest.

Instead of using the `text` member of the element, which only contains
the text up to the first subelement according to the ET documentation,
create proper HTML for the description.

All tags are transformed into `span` HTML tags with
`@class='doxygentag-ORIGINALTAGNAME` set to retain the original tag
information from the Doxygen schema. To give us a bit more semantic
HTML, the `<computeroutput>` tag is transformed into a `<code>` HTML tag.

Fixes #5001

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>
Signed-off-by: Philipp Wagner <phw@lowrisc.org>