[reggen] Define a class to represent the registers in a block
This is used to represent obj['registers'], which was previously a
list of registers, multiregs, windows, reserved and skipto entries.
The commit doesn't replace the top-level object (obj).

Other than removing genwennames and gennextoffset from (h)json export,
the only other change is that we no longer distinguish between
reserved and skipto, which means we have to pick one format or the
other on export. This commit picks reserved (because it is more
"relocatable"), but skipto could work instead.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>