[aes] Use sparse encodings for additional control signals
This commit extends the usage of sparse signal encodings to additional
control signals between main controller, counter module, cipher core
controller and inside the cipher core itself to protect the design from
starting or finishing early due to fault injection.

In case any of these signals takes on an invalid value, the corresponding
FSM signals an alert and enters the terminal error state. The design needs
to be reset.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>