[reggen] Fix layout in register table
I messed this up in commit 94926ae (by closing the <td> containing a
register description when there was an enum to follow).

This patch fixes that and, I think, simplifies the logic a bit. Now,
we collect up "desc_parts", which contains any register description
and any enum. The surrounding <td> ... </td> get added together at the
end. Doing it this way should avoid things coming unstuck again in the
future.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>