[prim_packer] Silence verilator width warnings
There are two changes here. The first is to explicitly zero-extend
wdata_i from InW up to OutW when computing wdata_shifted in pack mode.
The second is to slice out just the bits we need from MaxW / MinW when
comparing depth_q against it (otherwise Verilator complains that we're
comparing a small bitvector against a 32-bit int).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>