[ci] Install OpenOCD in CI
The edatools OBS repository now has a versioned build of OpenOCD,
similar to what I've done for Verilator. Use this build of OpenOCD in
CI. For end users of OpenTitan the installation guide recommends
self-compiling OpenOCD, which remains a viable option.

OpenOCD 0.11.0 is actually not release yet, and we're using 0.11.0-rc2
for the time being, with the final release expected to happen any day
now.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>