[sw, dif_pinmux] Introduce DIF Pinmux header
Please note that this is a reworked version of the following PRs:
https://github.com/lowRISC/opentitan/pull/2042 (PadCtrl)
https://github.com/lowRISC/opentitan/pull/1763 (Pinmux)

This is due to major changes in hardware - PadCtrl and Pinmux have
been merged into one (Pinmux).

Please note that all of the PadCtrl functionality was written by Greg
Chadwick, and was copied into this change with some minor naming
changes.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>