[dv] Fix VPD dumping
The VPD `dump` command returns a file id when run with the `-file
<filename>` switch (which is what we do as a first step). That by
itself does not start creating the dump file.

The `-add` command is what actually starts the actual dumping. Without
providing the file id to the `-add` switch, it assumes that a new file
is supposed to be created. For worksites that have access to the FSDM
dumper license, it defaults to dumping with FSDB (despite providing the
`-w vpd` switch to dvsim), else it defaults to vpd. This explains why it
works at lowRISC but bombs at Google / WD.

In this fix, we capture the file id returned by the `dump -file` command
and pass it to the subsequent `dump -add` commands.

While this did seem to fix the VPD dumping, I was unable to open it out of
the box using `gtkwave`. FWIU, the reason is this happens because we
enable dumping of MDAs. gtkwave runs the `vpd2vcd` utility underneath to
convert it to vcd, but that is missing the `+includemda` switch which
causes it to bomb. I am not sure how folks at LR are handling this at
the moment.

Fixes #5354.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>