[dv/otp_ctrl] fix regression status mismatch
This PR fix otp_ctrl check_fail sequence status error.
In scb we are trying to check status field `check_pending`, but if
timeout is set to very small value, when we read status, the check could
already finished due to timeout.
This PR fixes it by only checking status when check is done.

Signed-off-by: Cindy Chen <chencindy@google.com>