[sw/dif] Introduce CSRNG header
This API implements an interface for the CSRNG hardware.

The API follows the naming conventions used in NIST SP 800-90Ar1:

- Instantiate.
- Update
- Reseed
- Generate
- Uninstantiate

The seed used for instantiation can either be provided directly in
hardware by an entropy source module; or directly by software. See
`dif_csrng_instantiate()` and `dif_csrng_instantiate_source` for
more details.

The following sequence of operations are required to initialize
and generate criptographic entropy from the CSRNG module:

- `dif_csrng_init()`
- `dif_csrng_configure()`
- `dif_csrng_instantiate()`
- `dif_csrng_generate()`
- `dif_csrng_unistnantiate()`

The following functions can be used for reseed and update operations:

- `dif_csrng_reseed()`
- `dif_csrng_update()`

The following utility functions are available to poll the state of
the hardware:

- `dif_csrng_get_cmd_interface_status()`
- `dif_csrng_get_output_status()`

Please see the following documentation for more information:
https://docs.opentitan.org/hw/ip/csrng/doc/

Remaining work:

- Add error status interface.
- Add internal state control and debug interface.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>